Refactor: Move Hilt modules from core to data layer and clean unused dependencies
- Created :data:common and moved CommonModule out of :core:common.
- Created :data:camera:low-light and moved LowLightBoostModule out of :core:camera:low-light.
- Created :data:camera:postprocess and moved PostProcessModule out of :core:camera:postprocess.
- Moved :core:camera:low-light-playservices to :data:camera:low-light-playservices.
- Removed unused Hilt/kapt plugins and dependencies from :core:common, :core:settings, :core:camera:low-light, and :core:camera:postprocess.
- Removed unused hilt.navigation.compose dependency from :ui:components:capture and explicitly declared androidx.core:core-ktx.
- Updated all consumer module dependencies and import statements across app, feature, data, and core modules.
