+++ app/src/main/java/com/google/jetpackcamera/AppModule.kt
@@ -15,10 +15,10 @@
  */
 package com.google.jetpackcamera
 
-import com.google.jetpackcamera.core.common.DefaultCaptureModeOverride
-import com.google.jetpackcamera.core.common.DefaultFilePathGenerator
-import com.google.jetpackcamera.core.common.DefaultSaveMode
 import com.google.jetpackcamera.core.common.FilePathGenerator
+import com.google.jetpackcamera.data.common.DefaultCaptureModeOverride
+import com.google.jetpackcamera.data.common.DefaultFilePathGenerator
+import com.google.jetpackcamera.data.common.DefaultSaveMode
 import com.google.jetpackcamera.model.CaptureMode
 import com.google.jetpackcamera.model.SaveMode
 import dagger.Module
