+++ core/camera/low-light/build.gradle.kts
@@ -17,8 +17,6 @@
 plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
-    alias(libs.plugins.kotlin.kapt)
-    alias(libs.plugins.dagger.hilt.android)
 }
 
 android {
@@ -40,12 +38,5 @@ android {
 
 dependencies {
     implementation(libs.camera.core)
-    implementation(libs.dagger.hilt.android)
-    kapt(libs.dagger.hilt.compiler)
     implementation(libs.kotlinx.coroutines.core)
-}
-
-// Allow references to generated code
-kapt {
-    correctErrorTypes = true
 }
