+++ core/settings/build.gradle.kts
@@ -17,8 +17,6 @@
 plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
-    alias(libs.plugins.kotlin.kapt)
-    alias(libs.plugins.dagger.hilt.android)
 }
 
 android {
@@ -46,14 +44,6 @@ android {
 dependencies {
     implementation(libs.kotlinx.coroutines.core)
 
-    // Hilt
-    implementation(libs.dagger.hilt.android)
-    kapt(libs.dagger.hilt.compiler)
-
     // Domain models
     implementation(project(":core:model"))
-}
-
-kapt {
-    correctErrorTypes = true
 }
