+++ core/settings/datastore-prefs/build.gradle.kts
@@ -77,6 +77,7 @@ dependencies {
     // Access Model and Settings Interface
     implementation(project(":core:model"))
     implementation(project(":core:common"))
+    implementation(project(":data:common"))
     implementation(project(":core:settings"))
 }
