+++ data/camera/low-light-playservices/src/main/AndroidManifest.xml
