+++ data/camera/low-light-playservices/src/main/java/com/google/jetpackcamera/core/camera/lowlight/playservices/PlayServicesLowLightBoostEffectProvider.kt
