+++ data/camera/postprocess/src/main/java/com/google/jetpackcamera/data/camera/postprocess/PostProcessModule.kt
@@ -13,8 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.jetpackcamera.core.camera.postprocess
+package com.google.jetpackcamera.data.camera.postprocess
 
+import com.google.jetpackcamera.core.camera.postprocess.ImagePostProcessor
+import com.google.jetpackcamera.core.camera.postprocess.ImagePostProcessorFeatureKey
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
