+++ data/common/src/main/java/com/google/jetpackcamera/data/common/CommonModule.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2023 The Android Open Source Project
+ * Copyright (C) 2026 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.jetpackcamera.core.common
+package com.google.jetpackcamera.data.common
 
 import dagger.Module
 import dagger.Provides
@@ -52,9 +52,6 @@ class CommonModule {
 @Retention(AnnotationRetention.BINARY)
 annotation class DefaultFilePathGenerator
 
-/**
- * provides the default [CaptureMode] to override in [LocalSettingsRepository]
- */
 @Qualifier
 @Retention(AnnotationRetention.BINARY)
 annotation class DefaultCaptureModeOverride
